add class validations to bets and users

+++ src/bet/bet.controller.ts
@@ -6,9 +6,13 @@ import {
   Delete,
   Body,
   Param,
+  HttpException,
+  HttpStatus,
+  UsePipes,
 } from '@nestjs/common';
 import { BetService } from './bet.service';
 import { BetDTO } from './bet.dto';
+import { ValidationPipe } from '../shared/validation.pipe';
 
 @Controller('api/bet')
 export class BetController {
@@ -20,22 +24,36 @@ export class BetController {
   }
 
   @Post()
+  @UsePipes(new ValidationPipe())
   create(@Body() data: BetDTO) {
     return this.betService.create(data);
   }
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.betService.findOne(id);
+    const bet = this.betService.findOne(id);
+    if (!bet) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+    return bet;
   }
 
   @Put(':id')
   update(@Param('id') id: string, @Body() data: Partial<BetDTO>) {
+    const bet = this.betService.findOne(id);
+    if (!bet) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
     return this.betService.update(id, data);
   }
 
   @Delete(':id')
   destroy(@Param('id') id: string) {
-    return this.betService.destroy(id);
+    const bet = this.betService.findOne(id);
+    if (!bet) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+    this.betService.destroy(id);
+    return bet;
   }
 }
